Add a new KeybindSettings value to allow empty keybinds
This removes the need to hard-code checking isValid() == false ||
isKeybindHeld() to allow empty keybinds for certain features.
Now the user can select whether or not they want to allow a feature to be
active without having to hold any keys for it.

Note that this only affects the isKeybindHeld() return value.
Empty keybinds won't trigger any callbacks.