Add a KeyCallbackAdjustable hotkey callback class
This class is meant to be used for hotkeys that also have mouse scrolling
functionality tied to them, to adjust some kind of value.

Those hotkeys are set to trigger on both key event edges, and the value
can be adjusted by scrolling with a mouse wheel while holding the hotkey
down, and then on release the boolean config will get toggled, IF the
value was NOT adjusted.

The scrolling part should be implemented in the mods.
This class handles the hotkey activation base logic.
The scrolling handler should call KeyCallbackAdjustable.setValueChanged()
to mark when a value was changed, to prevent the boolean config toggle
when the hotkey is released after that.

Another key callback can be provided to handle the value toggling,
and to print whatever custom message might be required.