Improve the hover text renderer positioning a bit
Changed it so that it moves the text completely to the left of the cursor,
if it can't fit on the right, but can fit on the left.
If it can't fit on either side of the cursor, then it's moved to touching
the edge of the screen that the cursor is closest to.