Add a search bar to the string list edit screen
To make the edit actions target the correct entry in the list,
and also to show the correct indices when filtering the list,
the data list widget now tracks the indices of filtered entries.

Note: If the filtered entries are sorted, then the indices are currently wrong
as the index list is not sorted according to the entry list sorting.