A bunch of fixes and improvements to the InfoOverlay and InfoArea etc.
Especially the StringListRendererWidget is becoming more usable,
and it's now handled by the InfoWidgetManager.

The InfoRendererWidgets now can also hold "markers", so that mods
can know which instance of a widget they were last using, so that they
don't create new instances if the previous one they used still exists.
This is mainly meant for finding the existing and configured widget
after game restarts.