Configs: Refactor some of the config options and value overriding
The override stuff is now in the BaseGenericConfig class instead
of just in the BooleanConfig class.

Note that all configs don't support the overrides at this point,
and it's quite unlikely that a server would want to override
something like String of File configs or some of the list configs...
But still, they should probably also be refactored to support the
overrides at some point so that it's a truly generic feature.