Refactor ConfigSearchInfo and ConfigOverideUtils
This change allows to relax the boolean config type requirement
for the search info.

The override refactor is additionally preparing for the support
to override other configs than boolean configs. In this commit that
support doesn't fully exist yet, and there is just a temporary hack
in place for boolean configs.