DataListEntrySelectionHandler: Some changes and improvements...
... to be able to better select exactly the entries you want,
regardless of if there are multiple identical entries.

This still won't work properly for "persistent selections",
especially if the screen allows sorting or filtering the list.
For that to work properly without nasty selection update hacks
whenever the list gets re-ordered or filtered, the entire data list
probably needs to be of selection wrapper objects to persistently
track the selection status for each entry.