MenuWidget/MenuEntryWidget: Change/simplify how closing the menu works
The MenuWidget is now responsible for closing the menu, and it doesn't
add the menu close hook to the entry widgets anymore.
Instead, with the recent mouse input handling changes, the click
handling can now be passed from the entry widget to the MenuWidget
under it to close the menu.

To close the menu after clicking on an entry, the MenuEntryWidget
returns `false` to indicate that it did not consume the click.
Thus the handling will then be passed along to the next widget
by "handling priority", and by default the priority is directly
the widget's z-index i.e. height/stacking position on the screen.