Rewrite all rendering stuff to use a custom VertexBuilder interface
Also refactored/cleaned up the BaseOverlayRenderer stuff somewhat.

Using this custom VertexBuilder interface cuts out a lot of noise
from porting to other mod loaders or MC versions when things keep
changing around in vanilla.

This current VanillaWrappingVertexBuilder may or may not be a temporary
solution just to wrap the vanilla class inside basically one class.
At least still having the BufferBuilder.State class in the interface
is not ideal. But it being an inner class it wasn't trivial
(without Mixin hacks) to replace it with a custom class and still
being able to feed the data back to the vanilla BufferBuilder.